remove unused Impl layer
interfaces are used only for one instance and dont provide an advantage for testing
